requirements.txt
alembic==1.0.6
Click==7.0
Django==2.1.5
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
image==1.5.27
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
numpy==1.16.0
Pillow==5.4.1
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.9
six==1.12.0
SQLAlchemy==1.2.16
torch==0.4.1
torchvision==0.2.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
